requirements.txt
torch==1.6.0
scipy==1.6.3
numba==0.53.1
pycocotools==2.0.2
scikit_image==0.18.1
matplotlib==3.4.2
opencv_python==4.5.2.54
fvcore==0.1.5
seaborn==0.11.1
tqdm==4.61.0
easydict==1.9
torchvision==0.7.0
numpy==1.20.2
fire==0.4.0
mayavi==4.7.3
Pillow==8.3.2
PyYAML==5.4.1
scikit_learn==0.24.2
skimage==0.0
tensorboardX==2.4
